giant commit--implement baby compiler

+++ neve/include/obj.h
@@ -43,6 +43,6 @@ void printObj(Val val);
 
 void freeObj(Obj *obj);
 
-size_t objAsStr(const char *buffer, Obj *obj);
+uint32_t objAsStr(const char *buffer, Obj *obj);
 
 #endif
