+++ neve/include/val.h
@@ -50,6 +50,6 @@ void freeValArr(ValArr *arr);
 void printVal(Val val);
 
 bool valsEq(Val a, Val b);
-size_t valAsStr(char *buffer, Val val);
+uint32_t valAsStr(char *buffer, Val val);
 
 #endif
